SSML code for the poem
Worked on this quite awhile, had some issues with nesting some of the elements (for instance I tried to apply the <prosody pitch="low" rate="medium"></prosody> to the whole text, but for some reason it didn't work and I had to add it to each utterance separately). For some reason I also didn't hear any difference in the volume when adjusting it (for example I wanted to adjust the volume of the word "dirt" with <prosody volume="+20dB"></prosody>). Also some issues with emphasizing words, would be great if we could go through these elements once again? :)